0.94.3

+++ web/admin/ua/prj_export.jsp
@@ -9,12 +9,17 @@
 	java.util.*"%><%
 if(!Convert.checkReqEmpty(request, out, "id"))
 	return;
+
 String id = request.getParameter("id") ;
+UAPrj p = UAManager.getInstance().getPrjById(id);
+if(p==null)
+	return;
+
 
 File fout =UAManager.getInstance().exportPrjToTmp(id) ;
 try(FileInputStream fis = new FileInputStream(fout);)
 {
-	WebRes.renderFile(response, fout.getName(), fis) ;
+	WebRes.renderFile(response, "prj_"+p.getName(), fis) ;
 }
 
 fout.delete();
